Add files via upload
TRAINING AND PLACEMENT PORTAL is a website developed in the windows platform for the training and placement department of the college in order to provide the details of its students in a database for the companies to their process of recruitment provided with a proper login. Technology used to build the website is PHP with Codeigniter.